refactor(satellite): move cdn and storage within a same package
