+++ src/libs/satellite/src/api/config.rs
@@ -1,3 +1,6 @@
+use crate::assets::storage::store::{
+    get_config_store as get_storage_config_store, set_config_store as set_storage_config_store,
+};
 use crate::auth::store::{
     get_config as get_auth_config_store, set_config as set_auth_config_store,
 };
@@ -6,9 +9,6 @@ use crate::db::store::{
     get_config_store as get_db_config_store, set_config_store as set_db_config_store,
 };
 use crate::db::types::config::DbConfig;
-use crate::storage::store::{
-    get_config_store as get_storage_config_store, set_config_store as set_storage_config_store,
-};
 use crate::types::interface::Config;
 use ic_cdk::trap;
 use junobuild_storage::types::config::StorageConfig;
