+++ src/libs/satellite/src/assets/cdn/constants.rs
