+++ src/libs/satellite/src/assets/cdn/helpers/mod.rs
