+++ src/libs/satellite/src/assets/cdn/mod.rs
