+++ src/libs/satellite/src/assets/cdn/strategies_impls/cdn.rs
