+++ src/libs/satellite/src/assets/cdn/strategies_impls/mod.rs
@@ -1,0 +1,2 @@
+pub mod cdn;
+pub mod storage;
