+++ src/libs/satellite/src/assets/mod.rs
