+++ src/libs/satellite/src/assets/storage/certified_assets/mod.rs
