+++ src/libs/satellite/src/assets/storage/certified_assets/runtime.rs
@@ -1,5 +1,5 @@
+use crate::assets::storage::strategy_impls::StorageState;
 use crate::memory::internal::STATE;
-use crate::storage::strategy_impls::StorageState;
 use crate::types::state::State;
 use junobuild_storage::certification::types::certified::CertifiedAssetHashes;
 use junobuild_storage::certified_assets::extend_and_init_certified_assets;
