+++ src/libs/satellite/src/assets/storage/certified_assets/upgrade.rs
@@ -1,4 +1,4 @@
-use crate::storage::certified_assets::runtime::init_certified_assets as init_runtime_certified_assets;
+use crate::assets::storage::certified_assets::runtime::init_certified_assets as init_runtime_certified_assets;
 use ic_cdk::spawn;
 use ic_cdk_timers::set_timer;
 use std::time::Duration;
