+++ src/libs/satellite/src/assets/storage/store.rs
@@ -1,10 +1,8 @@
-use crate::controllers::store::get_controllers;
-use crate::memory::internal::STATE;
-use crate::storage::assert::{
+use crate::assets::storage::assert::{
     assert_create_batch, assert_delete_asset, assert_get_asset, assert_list_assets,
 };
-use crate::storage::certified_assets::runtime::init_certified_assets as init_runtime_certified_assets;
-use crate::storage::state::{
+use crate::assets::storage::certified_assets::runtime::init_certified_assets as init_runtime_certified_assets;
+use crate::assets::storage::state::{
     count_assets_stable, delete_asset as delete_state_asset, delete_domain as delete_state_domain,
     get_asset as get_state_asset, get_assets_stable, get_config as get_state_config, get_config,
     get_content_chunks as get_state_content_chunks, get_domain as get_state_domain,
@@ -12,7 +10,9 @@ use crate::storage::state::{
     get_rule as get_state_rule, insert_config as insert_state_config,
     insert_domain as insert_state_domain,
 };
-use crate::storage::strategy_impls::{StorageAssertions, StorageState, StorageUpload};
+use crate::assets::storage::strategy_impls::{StorageAssertions, StorageState, StorageUpload};
+use crate::controllers::store::get_controllers;
+use crate::memory::internal::STATE;
 use crate::types::store::StoreContext;
 use candid::Principal;
 use junobuild_collections::msg::msg_storage_collection_not_empty;
