+++ src/libs/satellite/src/assets/storage/strategy_impls.rs
@@ -1,10 +1,10 @@
-use crate::cdn::assert::assert_cdn_asset_keys;
-use crate::hooks::storage::invoke_assert_upload_asset;
-use crate::storage::assert::assert_storage_list_permission;
-use crate::storage::state::{
+use crate::assets::cdn::assert::assert_cdn_asset_keys;
+use crate::assets::storage::assert::assert_storage_list_permission;
+use crate::assets::storage::state::{
     delete_asset, get_asset, get_config, get_domains, get_rule, insert_asset, insert_asset_encoding,
 };
-use crate::storage::store::{get_content_chunks_store, get_public_asset_store};
+use crate::assets::storage::store::{get_content_chunks_store, get_public_asset_store};
+use crate::hooks::storage::invoke_assert_upload_asset;
 use crate::user::usage::assert::increment_and_assert_storage_usage;
 use candid::Principal;
 use junobuild_collections::assert::stores::{assert_create_permission, assert_permission};
