+++ src/libs/satellite/src/assets/storage/types.rs
