+++ src/libs/satellite/src/memory/lifecycle.rs
@@ -1,3 +1,4 @@
+use crate::assets::storage::certified_assets::upgrade::defer_init_certified_assets;
 use crate::hooks::lifecycle::{
     invoke_on_init, invoke_on_init_sync, invoke_on_post_upgrade, invoke_on_post_upgrade_sync,
 };
@@ -5,7 +6,6 @@ use crate::memory::internal::{get_memory_for_upgrade, init_stable_state, STATE};
 use crate::memory::upgrade::init_juno_collection;
 use crate::memory::utils::init_storage_heap_state;
 use crate::random::init::defer_init_random_seed;
-use crate::storage::certified_assets::upgrade::defer_init_certified_assets;
 use crate::types::state::{HeapState, RuntimeState, State};
 use ciborium::{from_reader, into_writer};
 use ic_cdk::api::call::{arg_data, ArgDecoderConfig};
