+++ src/libs/satellite/src/memory/upgrade.rs
@@ -1,4 +1,4 @@
-use crate::cdn::constants::{
+use crate::assets::cdn::constants::{
     CDN_JUNO_RELEASES_COLLECTION_KEY, DEFAULT_CDN_JUNO_RELEASES_COLLECTIONS,
 };
 use crate::memory::internal::STATE;
