+++ src/libs/satellite/src/sdk/core/storage.rs
@@ -1,5 +1,5 @@
-pub use crate::storage::handlers::set_asset_handler;
-pub use crate::storage::store::{
+pub use crate::assets::storage::handlers::set_asset_handler;
+pub use crate::assets::storage::store::{
     count_assets_store, count_collection_assets_store, delete_asset_store, delete_assets_store,
     delete_filtered_assets_store, get_asset_store, get_content_chunks_store, list_assets_store,
 };
